+++ include/Config.hpp
@@ -1,4 +1,0 @@
-// the configured options and settings for Tutorial
-#define RNAnue_VERSION_MAJOR 0
-#define RNAnue_VERSION_MINOR 2
-#define RNAnue_VERSION_PATCH 1