+++ src/main.cpp
@@ -198,7 +198,6 @@ int main(int argc, char* argv[]) {
                 notify(params);
             }
         }
-
         Base base(params);
 
     } catch(po::error& e) {