+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [0.2.3]
+
+## Fix
+
+- Consumer/Producer pattern for 'detect' to integrate concurrency
+- Write to .fq using SeqAn3 write operations
+- Assertion error when writing to .fq (caused by variable initialization outside worker)
+
+
 # [0.2.2]
 
 ## Features